requirements.txt
requests>=2.31
python-dotenv>=1.0
Pillow>=10.0
fal-client>=0.5.0
